# openapi.yml
openapi: 3.0.0
info:
  title: Gironeta Server API
  description: This file defines the communication contract between the AI league application front-end and the back-end service.
  version: 1.0.0
paths:
...
  /api/v0/league:
    post:
      summary: Create a new league
      description: This endpoint allows a user to create a new league.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeagueDTO'
      responses:
        '200':
          description: League created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDTO'
    get:
      summary: Get leagues
      description: This endpoint retrieves leagues, allowing to filter results by owner.
      parameters:
        - name: owner
          in: query
          description: Filter leagues by owner
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of leagues retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeagueDTO'
  /api/v0/league/{leagueId}:
    get:
      summary: Get league by ID
      description: This endpoint retrieves a league by its ID.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: League retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDTO'
        '404':
          description: League not found
    put:
      summary: Update league by ID
      description: This endpoint updates a league by its ID.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeagueDTO'
      responses:
        '200':
          description: League updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDTO'
        '404':
          description: League not found
    delete:
      summary: Delete league by ID
      description: This endpoint deletes a league by its ID.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to delete
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: League deleted successfully
        '404':
          description: League not found
  /api/v0/league/{leagueId}/bot:
    post:
      summary: Add bots to a league
      description: This endpoint allows a user to add bots to a league.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to which the bot will be added
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              description: List of bot IDs to add to the league
              items:
                type: integer
                format: int64
      responses:
        '204':
          description: Bots added to league successfully
        '404':
          description: League not found
        '400':
          description: Bad request, invalid bot IDs
  /api/v0/league/{leagueId}/leaderboard:
    get:
      summary: Get league leaderboard
      description: This endpoint retrieves the leaderboard of a league.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to retrieve the leaderboard for
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: League leaderboard retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipationDTO'
        '404':
          description: League not found
  /api/v0/league/{leagueId}/start:
    post:
      summary: Start a league
      description: This endpoint starts a league.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to start
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: League started successfully
        '404':
          description: League not found
  /api/v0/league/{leagueId}/match:
    get:
      summary: Get matches of a league
      description: This endpoint retrieves the matches of a league.
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the league to retrieve matches for
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: League matches retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDTO'
        '404':
          description: League not found
...
components:
  schemas:
...
    BotDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        quality:
          type: string
        imageUrl:
          type: string
        apiUrl:
          type: string
        nWins:
          type: integer
        nDraws:
          type: integer
        nLosses:
          type: integer
      required:
        - id
        - name
        - quality
        - apiUrl
...
    CreateLeagueDTO:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
        rounds:
          type: integer
        matchTime:
          type: integer
          format: int64
      required:
        - name
        - rounds
        - matchTime
    LeagueDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        imageUrl:
          type: string
        rounds:
          type: integer
        matchTime:
          type: integer
          format: int64
        bots:
          type: array
          items:
            $ref: '#/components/schemas/BotDTO'
        state:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
      required:
        - id
        - name
        - rounds
        - matchTime
        - bots
        - state
    ParticipationDTO:
      type: object
      properties:
        botId:
          type: integer
          format: int64
        botName:
          type: string
        points:
          type: integer
        position:
          type: integer
        nWins:
          type: integer
        nDraws:
          type: integer
        nLosses:
          type: integer
      required:
        - botId
        - botName
        - points
        - position
        - nWins
        - nDraws
        - nLosses
    MatchDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        state:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
        result:
          type: integer
          description: Result of the match, -1 for draw, otherwise the winner's index in the fighters array
        fighters:
          type: array
          items:
            $ref: '#/components/schemas/BotDTO'
        roundNumber:
          type: integer
      required:
        - id
        - state
        - result
        - fighters
        - roundNumber
...